+++ lib/src/beeper/beeper_channel_contract.dart
@@ -1,0 +1,4 @@
+class BeeperChannelContract {
+  static const String beeperBeep = 'beeperBeep';
+  static const String beeperStop = 'beeperStop';
+}