+++ lib/src/device/device_channel_contract.dart
@@ -1,0 +1,3 @@
+class DeviceChannelContract {
+  static const String deviceGetStatus = 'deviceGetStatus';
+}